Refactor dashboard (#265)
